Refactor signer resolution and bitmap handling in verification logic
- Updated `README.md` to clarify function usage for signer resolution.
- Simplified the `resolve_signers` function and removed the indexed variant.
- Enhanced bitmap handling in `bitmap.rs` and adjusted related functions for clarity.
- Refactored `onchain.rs` and `verify.rs` to utilize the updated signer resolution logic.
- Updated tests to ensure compatibility with the refactored verification and bitmap handling.

+++ src/bitmap.rs
@@ -130,7 +130,7 @@ where
     F: FnMut(usize) -> Result<(), E>,
 {
     while bm != Bitmap::EMPTY {
-        let bit_pos = bm.trailing_zeros() as usize;
+        let bit_pos = bm.trailing_zeros();
         bm.clear_bit(bit_pos);
         f(bit_pos)?;
     }
